+++ CHANGELOG.rst
@@ -2,6 +2,15 @@
 Changelog
 =========
 
+Version 0.0.2 (unreleased)
+===========================
+
+Bugfixes
+--------
+
+- Fixed issue where empty or very fast-running scripts would not generate performance charts
+- Performance charts are now always generated, even when script execution time is too short to collect data
+
 Version 0.0.1 (2025-11-03)
 ==========================